Nestable Networks validation and modules reorganization
