+++ tensor-safe/src/TensorSafe/Layers.hs
@@ -1,0 +1,17 @@
+module TensorSafe.Layers (
+    Conv2D,
+    Dense,
+    Dropout,
+    Flatten,
+    MaxPooling,
+    Relu,
+    Sigmoid
+) where
+
+import           TensorSafe.Layers.Conv2D
+import           TensorSafe.Layers.Dense
+import           TensorSafe.Layers.Dropout
+import           TensorSafe.Layers.Flatten
+import           TensorSafe.Layers.MaxPooling
+import           TensorSafe.Layers.Relu
+import           TensorSafe.Layers.Sigmoid
